requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
colorama==0.4.4
distlib==0.3.4
Django==4.0.1
djangorestframework==3.13.1
filelock==3.4.2
iniconfig==1.1.1
packaging==21.3
pipenv==2022.1.8
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
virtualenv==20.13.0
virtualenv-clone==0.5.7
